GET working, Post Working.
----
Needs verification and bonus marks.

+++ src/com/COMP6461/server/SocketConnectionFromServer.java
@@ -3,6 +3,8 @@ package com.COMP6461.server;
 import java.io.*;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Date;
 import java.util.StringTokenizer;
 
@@ -43,31 +45,42 @@ public class SocketConnectionFromServer implements Runnable {
     public void run() {
         BufferedReader in = null;
         this.serverRequest = new HttpServerRequest();
+        System.out.println("Server started with request from client...");
         try {
             in = new BufferedReader(new InputStreamReader(this.socket.getInputStream()));
             String input = in.readLine();
             StringTokenizer parse = new StringTokenizer(input);
             this.serverRequest.method = parse.nextToken().toUpperCase(); // we get the HTTP method of the client
             this.serverRequest.uri = parse.nextToken().toLowerCase();
-            //"E:\\My stuff\\Masters Study Material\\6th term\\Assignment2\\serverDirectory\\abc.txt";
-            boolean readHeaders = true;
-            while ((input = in.readLine()) != null) {
-                if (input == "") {
-                    readHeaders = false;
+            int postDataI = -1;
+            String line = "";
+            while ((line = in.readLine()) != null && (line.length() != 0)) {
+                this.serverRequest.headers.add(line);
+                if (line.indexOf("Content-Length:") > -1) {
+                    postDataI = new Integer(
+                            line.substring(
+                                    line.indexOf("Content-Length:") + 16,
+                                    line.length())).intValue();
                 }
-                if (readHeaders) {
-                    this.serverRequest.headers.add(input);
-                } else {
-                    this.serverRequest.body.append(input);
-                }
-                this.performOperation();
             }
+            String postData = "";
+            // read the post data
+            if (postDataI > 0) {
+                char[] charArray = new char[postDataI];
+                in.read(charArray, 0, postDataI);
+                postData = new String(charArray);
+            }
+            System.out.println("postdataad : \n " + postData);
+            this.serverRequest.body.append(postData);
+            this.performOperation();
+            System.out.println(this.serverRequest);
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
     public void performOperation() {
+        System.out.println("Operation started");
         responseToSend = new StringBuilder();
         responseToSend.append("HTTP/1.1 ");
         if (this.serverRequest.method.toLowerCase().equals("get")) {
@@ -77,7 +90,8 @@ public class SocketConnectionFromServer implements Runnable {
                 responseToSend.append("\nDate: " + new Date());
                 responseToSend.append("\n" + SERVER_NAME);
                 StringBuilder listOfFilesNames = new StringBuilder();
-                File folder = new File(getClass().getResource("abc.txt").getPath());
+                Path currentRelativePath = Paths.get("");
+                File folder = new File(currentRelativePath.toAbsolutePath().toString());
                 File[] listOfFiles = folder.listFiles();
                 for (int i = 0; i < listOfFiles.length; i++) {
                     System.out.println("File " + listOfFiles[i].getName());
@@ -86,11 +100,12 @@ public class SocketConnectionFromServer implements Runnable {
                 }
                 responseToSend.append("\nContent-Length:" + listOfFilesNames.length());
                 responseToSend.append("\n");
-                responseToSend.append(listOfFilesNames);
+                responseToSend.append("\n{" + listOfFilesNames + "\n}");
 
 
             } else {
-                File file = new File(this.serverRequest.uri);
+                Path currentRelativePath = Paths.get("");
+                File file = new File(currentRelativePath.toAbsolutePath().toString() + this.serverRequest.uri);
                 if (!file.exists()) {
                     System.out.println("Wrong folder 404");
                     responseToSend.append("404 Not Found");
@@ -108,8 +123,8 @@ public class SocketConnectionFromServer implements Runnable {
                         bufferedReader = new BufferedReader(new FileReader(file));
                         while ((readLine = bufferedReader.readLine()) != null) {
                             System.out.println(readLine);
-                            linesFromFile.append(readLine);
-                            responseToSend.append("\n " + readLine);
+                            linesFromFile.append("\n" + readLine);
+                            // responseToSend.append("\n " + readLine);
                         }
                         bufferedReader.close();
                     } catch (IOException e) {
@@ -118,33 +133,47 @@ public class SocketConnectionFromServer implements Runnable {
                     responseToSend.append("\nContent-type: text/plain");
                     responseToSend.append("\nContent-Length:" + linesFromFile.length());
                     responseToSend.append("\n");
-                    responseToSend.append(linesFromFile);
+                    responseToSend.append("\n{" + linesFromFile + "\n}");
                 }
             }
         }
         if (this.serverRequest.method.toLowerCase().equals("post")) {
             if (!this.serverRequest.uri.equals("")) {
-                File file = new File(this.serverRequest.uri);
-                try {
-                    BufferedWriter writer = new BufferedWriter(new FileWriter(file, false));
-                    writer.write(this.serverRequest.body.toString());
-                    responseToSend.append("200 Ok");
+                Path currentRelativePath = Paths.get("");
+                File file = new File(currentRelativePath.toAbsolutePath().toString() + this.serverRequest.uri);
+                if (file.exists()) {
+                    try {
+                        BufferedWriter writer = new BufferedWriter(new FileWriter(file, false));
+                        writer.write(this.serverRequest.body.toString());
+                        responseToSend.append("200 Ok");
+                        responseToSend.append("\nDate: " + new Date());
+                        responseToSend.append("\n" + SERVER_NAME);
+                        responseToSend.append("\n");
+                        responseToSend.append("\n{" + this.serverRequest.body + "\n}");
+                        writer.close();
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                } else {
+                    responseToSend.append("404 File not Found");
                     responseToSend.append("\nDate: " + new Date());
                     responseToSend.append("\n" + SERVER_NAME);
-                    writer.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
+                    responseToSend.append("\n");
+                    responseToSend.append("\n{" + this.serverRequest.body + "\n}");
                 }
-
             }
         }
         this.sendResponse();
     }
 
     public void sendResponse() {
+        System.out.println(this.responseToSend);
         try {
             PrintWriter out = new PrintWriter(socket.getOutputStream());
             out.print(this.responseToSend);
+            out.flush();
+            out.close();
+            socket.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
